Subquery: Subqueries can now use custom semirings & evaluators
This adds a new subquery/5 and subquery_in_scope/6 that allows setting a
custom semiring and custom evaluator. To support this a semiring registry
was added. Custom semirings can be registered with it (see
problog/__init__.py).

Semirings registered with this registry needs a class method create(), that
given a database, an engine, and various keyword arguments sent to the
builtin constructs a new instance.